Add files via upload
Batch file to minidump (with mimikatz) .dmp files in the current directory automatically. The .dmp fils should be named uniquely. Results will be written to an 'output.txt' file and previously dumped .dmp files will be added to a 'DUMPED' folder (created upon initial execution. Additional dumps will be appended to the 'output.txt' file and also added to the 'DUMPED' folder for tracking.